Rename/reorganize as per review
